+++ NEWS.md
@@ -1,5 +1,16 @@
 # History of user-visible changes
 
+## 2020-12-20
+
+* Support for logical assignment operators ([#564](https://github.com/mooz/js2-mode/issues/564)).
+* Support for the nullish coalescing operator ([#561](https://github.com/mooz/js2-mode/pull/561)).
+* Emacs 27 now provides improved JSX indentation support, along with
+  new JSX highlighting and detection support. Install Emacs 27 and use
+  `js-mode` with `js2-minor-mode` (see README), rather than
+  `js2-jsx-mode`.
+* Using `js2-jsx-mode` will now trigger a warning in Emacs 27.
+* Support for optional-chaining operator `?.`
+
 ## 2019-02-19
 
 * Changed the default of `js2-strict-trailing-comma-warning` to nil.